Revert "Restructures src directory and implements RollUpJS (#128)"
This reverts commit 749173c0b97cab1f2a654d45dee1758780ef9e3d.
