+++ .travis.yml
@@ -6,6 +6,6 @@ dist: bionic
 install:
  - npm install
  - npm install -g grunt-cli
- - grunt clean copy rollup
+ - grunt clean copy
 script:
  - npm test
