+++ package.json
@@ -46,14 +46,12 @@
     "grunt-contrib-uglify-es": "~3.3.0",
     "grunt-contrib-watch": "~1.1.0",
     "grunt-html": "~9.3.0",
-    "grunt-rollup": "^11.3.0",
     "jest": "^26.1.0",
     "jest-playwright-preset": "^0.1.3",
     "leaflet": "~1.6.0",
     "path": "^0.12.7",
     "playwright": "^1.2.1",
     "proj4": "2.6.0",
-    "proj4leaflet": "1.0.1",
-    "rollup": "^2.23.1"
+    "proj4leaflet": "1.0.1"
   }
 }
