--- rollup.config.js
@@ -1,7 +1,0 @@
-export default {
-  input: 'src/index.js',
-  output: {
-    file: 'bundleTest/mapml.js',
-    format: 'iife',
-  },
-};
