--- src/mapml/utils/Constants.js
@@ -1,3 +1,0 @@
-export const TILE_SIZE = 256;
-export const FALLBACK_PROJECTION = "OSMTILE";
-export const FALLBACK_CS = "TILEMATRIX";
