+++ test/server.js
@@ -3,8 +3,9 @@ const app = express();
 const path = require("path");
 const port = 30001;
 
+//loads in the src first
+app.use(express.static(path.join(__dirname, "../src")));
 //then loads in the index file
-app.use(express.static(path.join(__dirname, "../dist")));
 app.use(express.static(path.join(__dirname, "e2e/core")));
 app.use(express.static(path.join(__dirname, "e2e/layers")));
 app.use("/data", express.static(path.join(__dirname, "e2e/data/tiles/cbmt")));
@@ -30,6 +31,8 @@ app.use(
   express.static(path.join(__dirname, "e2e/data/tiles/wgs84/1"))
 );
 
+//lastly loads the dist, but the original src files have priority
+app.use(express.static(path.join(__dirname, "../dist")));
 console.log("Running on localhost:" + port);
 
 app.listen(port);
