+++ test/setup.js
@@ -1,7 +1,7 @@
 require("../dist/leaflet-src.js");
 require("../dist/proj4-src.js");
 require("../dist/proj4leaflet.js");
-require("../dist/mapml.js");
+require("../src/mapml.js");
 
 /* 'dist/leaflet-src.js',
       'dist/proj4-src.js',
